requirements_backend.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
numpy>=1.24.0

